pyproject.toml
[tool.poetry]
name = "liboard"
version = "2.0.3"
description = "A module for interacting with LiBoard"
authors = ["Philipp Leclercq <liboard@mailbox.org>"]
license = "GPL-3.0-only"

[tool.poetry.dependencies]
python = "^3.9"
python-chess = "^1.999"
pyserial = "^3.5"
bitstring = "^3.1.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
